#164 refactor the test class
